3个异步回调+退款

See merge request langxiuchen/ooad201!29

+++ transaction/src/main/java/cn/edu/xmu/oomall/transaction/model/bo/RefundState.java
@@ -1,19 +1,19 @@
 package cn.edu.xmu.oomall.transaction.model.bo;
 
 public enum RefundState {
-    WAIT_REFUND(0,"待退款"),
-    FINISH_REFUND(1,"已退款"),
-    FINISH_RECONCILIATION(2,"已对账"),
-    FINISH_LIQUIDATION(3,"已清算"),
-    CANCEL_REFUND(4,"已取消"),
-    FAILED(5,"失败");
-    private int code;
+    WAIT_REFUND((byte)0,"待退款"),
+    FINISH_REFUND((byte)1,"已退款"),
+    FINISH_RECONCILIATION((byte)2,"已对账"),
+    FINISH_LIQUIDATION((byte)3,"已清算"),
+    CANCEL_REFUND((byte)4,"已取消"),
+    FAILED((byte)5,"失败");
+    private Byte code;
     private String message;
-    RefundState(int code, String message){
+    RefundState(Byte code, String message){
         this.code = code;
         this.message = message;
     }
-    public int getCode() {
+    public Byte getCode() {
         return code;
     }
